move getTag() to other file
